+++ firstSolar/back/package.json
@@ -12,6 +12,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@moralisweb3/common-evm-utils": "^2.18.2",
     "@openzeppelin/contracts": "^4.8.2",
     "@pinata/sdk": "^2.1.0",
     "@remix-project/remixd": "^0.6.13",
