+++ firstSolar/front/webpack.config.js
@@ -1,0 +1,10 @@
+import { resolve } from "path";
+import { require } from "module";
+
+export default {
+  resolve: {
+    fallback: {
+      stream: require.resolve("stream-browserify"),
+    },
+  },
+};
